+++ src/main/kotlin/klib/objects/library/Library.kt
@@ -3,6 +3,7 @@ package klib.objects.library
 import klib.types.library.LClass
 import klib.types.library.LFunction
 import java.io.File
+import java.io.FileNotFoundException
 import java.net.URL
 import java.net.URLClassLoader
 
@@ -22,13 +23,17 @@ object Library {
      * @param file The Jar file to load from
      * @param className The Class to load
      * @return Returns a LClass object
+     * @throws FileNotFoundException
+     * @throws ClassNotFoundException
      *
      * @see LClass
      *
      * @since 1.3.0 (Experimental)
      * @author Thomas Obernosterer
      */
+    @Throws(FileNotFoundException::class, ClassNotFoundException::class)
     fun loadClassFromJar(file: File, className: String): LClass {
+        if (!file.exists()) throw FileNotFoundException()
         val child = URLClassLoader(arrayOf<URL>(file.toURI().toURL()), this.javaClass.classLoader)
         val classFromJar = Class.forName(className, true, child)
         return LClass(classFromJar)
@@ -41,6 +46,9 @@ object Library {
      * @param className The Class to load from
      * @param functionName The Method to load
      * @return Returns a LFunction object
+     * @throws FileNotFoundException
+     * @throws ClassNotFoundException
+     * @throws NoSuchMethodException
      *
      * @see LClass
      * @see LFunction
@@ -48,10 +56,12 @@ object Library {
      * @since 1.3.0 (Experimental)
      * @author Thomas Obernosterer
      */
+    @Throws(FileNotFoundException::class, ClassNotFoundException::class, NoSuchMethodException::class)
     fun loadFunctionFromJar(file: File, className: String, functionName: String): LFunction {
+        if (!file.exists()) throw FileNotFoundException()
         val child = URLClassLoader(arrayOf<URL>(file.toURI().toURL()), this.javaClass.classLoader)
         val classFromJar = Class.forName(className, true, child)
-        val method = classFromJar.getMethod(functionName)
+        val method = classFromJar.methods.find { it.name == functionName } ?: throw NoSuchMethodException()
         val classObject = classFromJar.newInstance()
         return LFunction(classObject, method)
     }